fix: A seção AfterAll dos testes apresentava o erro de NullPointerException

+++ 03-conversor-csv/app/src/test/java/com/exercise/AppTest.java
@@ -105,16 +105,18 @@ class AppTest {
   @AfterAll
   public static void after() throws IOException {
 
-    for (File folder : pathTest.listFiles()) {
-      for (File file : folder.listFiles()) {
-        file.delete();
+    if (pathTest.exists()) {
+      for (File folder : pathTest.listFiles()) {
+        for (File file : folder.listFiles()) {
+          file.delete();
+        }
       }
+      
+      entriesFolder.delete();
+      outputFolder.delete();
+      
+      pathTest.delete();
     }
-
-    entriesFolder.delete();
-    outputFolder.delete();
-
-    pathTest.delete();
   }
 
   @Test
